rag/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.135.1
frozenlist==1.8.0
greenlet==3.3.2
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.0
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.20
langchain-openai==1.1.11
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.1
langsmith==0.7.22
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.3
ollama==0.6.1
openai==2.28.0
orjson==3.11.7
packaging==26.0
portalocker==3.2.0
propcache==0.4.1
protobuf==7.34.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pypdf==6.9.1
python-dotenv==1.2.2
PyYAML==6.0.3
qdrant-client==1.17.1
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.52.1
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
